edit tour and daybook

+++ src/main/java/com/example/capstone_be/dto/tour/TourByCategoryDto.java
@@ -21,7 +21,7 @@ public class TourByCategoryDto implements Serializable {
 
     private String city;
 
-    private Float priceOnePerson;
+    private Double priceOnePerson;
 
     private String imageMain;
