+++ src/main/java/com/example/capstone_be/dto/tour/TourDetailDto.java
@@ -23,7 +23,7 @@ public class TourDetailDto implements Serializable {
 
     private String city;
 
-    private Float priceOnePerson;
+    private Double priceOnePerson;
 
     private String imageMain;
