+++ src/main/java/com/example/capstone_be/dto/tour/TourViewDto.java
@@ -23,7 +23,7 @@ public class TourViewDto implements Serializable {
 
     private String city;
 
-    private Float priceOnePerson;
+    private Double priceOnePerson;
 
     private String imageMain;
